- Updated the download manager to utilize static buffers for various paths and keys, reducing stack usage and improving performance during downloads.
- Implemented input validation in the load tracker for path building, ensuring robustness against invalid arguments.
- Adjusted the heap allocation strategy in the load tracker to minimize stack usage during deep call chains.
- Enhanced the play scheduler with new functions to retrieve channel entry counts and check for Makapix channels, streamlining download management.

These changes improve the efficiency and reliability of the download and channel management processes in the p3a system.

+++ components/play_scheduler/include/play_scheduler_types.h
@@ -202,7 +202,7 @@ typedef struct {
 
     // Pick state
     uint32_t cursor;          // For RecencyPick
-    uint32_t pick_rng_state;  // For RandomPick
+    uint64_t pick_rng_state;  // For RandomPick (64-bit state for proper PCG32)
 
     // Cache info
     size_t entry_count;       // Mi: local cache size
